Fix racy uaf on join
When threads shutdown they unregister by decrementing the spawn count. The last to make it zero will set the `join_event` to begin the thread-kill-chain. 

The join side tried to avoid calling `join_event.wait()` by checking for spawn count of zero before-hand. There's a race where
1) last thread sets to 0 but is preempted before setting the event 2) join thread sees 0, skips waiting, and invalidates the pool/event 3) last thread resumes and sets an event being invalidated.

The fix should be to always wait for the last thread's `join_event.set` to ensure it will no longer have access to the pool when going to invalidate it.